Finish User/Admin validation in all routes

+++ controllers/reviewController.js
@@ -1,10 +1,12 @@
 import Review from "../models/reviewModel.js";
 import User from "../models/userModel.js";
-import Experience from "../models/experienceModel.js";
+import Admin from "../models/adminModel.js";
 
 async function listReview(req, res) {
   try {
-    const reviewList = await Review.find().populate();
+    const reviewList = await Review.find()
+      .populate("user")
+      .populate("experience");
     res.json(reviewList);
   } catch (error) {
     res.status(500).json("The server had an error");
@@ -37,25 +39,37 @@ async function findReview(req, res) {
 
 async function editReview(req, res) {
   try {
-    const {id} = await User.findById(req.auth.sub);
-    const reviewUserId = await Review.findById(req.params.user);
-
-    if( id === reviewUserId[0].toString()){
-
+    const userId = req.auth.sub;
+    const user = await User.findById(userId);
+    const idUser = user ? user.id : null;
+    const admin = await Admin.findById(userId);
+    const idAdmin = admin ? admin.id : null;
     const foundReview = await Review.findById(req.params.id);
 
-    foundReview.user = req.body.user ?? foundReview.user;
-    foundReview.experience = req.body.experience ?? foundReview.experience;
-    foundReview.comment = req.body.comment ?? foundReview.comment;
-    foundReview.score = req.body.score ?? foundReview.score;
+    if (idUser !== null) {
+      if (userId === foundReview.user[0].toString()) {
+        foundReview.user = req.body.user ?? foundReview.user;
+        foundReview.experience = req.body.experience ?? foundReview.experience;
+        foundReview.comment = req.body.comment ?? foundReview.comment;
+        foundReview.score = req.body.score ?? foundReview.score;
 
-    await foundReview.save();
+        await foundReview.save();
 
-    res.json(foundReview);
-  } else {
-    res.json ( "this is not your review, check again");
-  }
-  
+        res.json(foundReview);
+      } else {
+        res.json("this is not your review, check again");
+      }
+    }
+    if (idAdmin !== null) {
+      foundReview.user = req.body.user ?? foundReview.user;
+      foundReview.experience = req.body.experience ?? foundReview.experience;
+      foundReview.comment = req.body.comment ?? foundReview.comment;
+      foundReview.score = req.body.score ?? foundReview.score;
+
+      await foundReview.save();
+
+      res.json(foundReview);
+    }
   } catch (error) {
     res.status(500).json("The server had an error");
   }
@@ -63,15 +77,24 @@ async function editReview(req, res) {
 
 async function deleteReview(req, res) {
   try {
-    const {id} = await User.findById(req.auth.sub)
-    const reviewUserId = await Review.findById(req.params.id);
-    if ( id === reviewUserId ){
-
-    const deleteReview = await Review.findByIdAndDelete(req.params.id);
-    res.json("The review was deleted");
-  } else {
-    res.json("you cannot delete this review, check again");
-  }
+    const userId = req.auth.sub;
+    const user = await User.findById(userId);
+    const idUser = user ? user.id : null;
+    const admin = await Admin.findById(userId);
+    const idAdmin = admin ? admin.id : null;
+    const foundReview = await Review.findById(req.params.id);
+    if (idUser !== null) {
+      if (userId === foundReview.user[0].toString()) {
+        const deleteReview = await Review.findByIdAndDelete(req.params.id);
+        res.json("The review was deleted");
+      } else {
+        res.json("you cannot delete this review, check again");
+      }
+    }
+    if (idAdmin !== null) {
+      const deleteReview = await Review.findByIdAndDelete(req.params.id);
+      res.json("The review was deleted");
+    }
   } catch (error) {
     res.status(500).json("The server had an error");
   }
